+++ internal/httputil/responder.go
@@ -16,8 +16,8 @@ import (
 // HashGetter returns a freshly calculated or cached hash.
 type HashGetter func() (string, error)
 
-func NewResponder(w http.ResponseWriter, r *http.Request, contentType string, allowOrigin string) *Responder {
-	return &Responder{w, r, contentType, allowOrigin}
+func NewResponder(w http.ResponseWriter, r *http.Request, contentType string) *Responder {
+	return &Responder{w, r, contentType}
 }
 
 type Responder struct {
@@ -25,11 +25,6 @@ type Responder struct {
 	r *http.Request
 
 	ContentType string
-
-	// The value of the Access-Control-Allow-Origin HTTP header to set, if empty
-	// the header will remain unset. See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin
-	// for valid values.
-	allowOrigin string
 }
 
 func (re *Responder) Cached(etag HashGetter) bool {
@@ -55,10 +50,6 @@ func (re *Responder) Cache(etag HashGetter) {
 }
 
 func (re *Responder) OK(data interface{}) {
-	if re.allowOrigin != "" {
-		re.w.Header().Set("Access-Control-Allow-Origin", re.allowOrigin)
-		re.w.Header().Set("Vary", "Origin")
-	}
 	re.w.Header().Set("Content-Type", re.ContentType)
 
 	if re.ContentType != "application/json" {
@@ -80,10 +71,6 @@ func (re *Responder) OK(data interface{}) {
 func (re *Responder) Error(hErr *Error, err error) {
 	if hErr.Code != http.StatusNotFound {
 		log.Printf("Error (masked) while responding to %s: %s", re.r.URL, err)
-	}
-	if re.allowOrigin != "" {
-		re.w.Header().Set("Access-Control-Allow-Origin", re.allowOrigin)
-		re.w.Header().Set("Vary", "Origin")
 	}
 	re.w.Header().Set("Content-Type", re.ContentType)
 	re.w.WriteHeader(hErr.Code)